fix(insanetweaks) 1.9.12: read enchantment ids as int, not short
Sentient Codex resolved enchantment ids out of the `ench` NBT list with
getShort("id"). JustEnoughIDs widens enchantment ids from short to int and
rewrites vanilla's own reads to match, but it does not touch ours - so any id
above 32767 came back truncated. In EnchantmentSentientCodex.getSentientCodexLevel
that silently reports level 0 (the enchantment stops working); in
SentientCodexHandler it makes the boost loop skip or mis-identify entries.

Latent today - the pack is a few hundred enchantments deep, well under the
limit - so this is hardening, not a live fix.

NBTTagCompound.getInteger accepts any numeric tag, so a short-typed "id" still
reads correctly and the change is safe without JEID installed. "lvl" stays a
short deliberately: JEID widens only the id.

+++ insanetweaks/src/main/java/com/spege/insanetweaks/enchant/EnchantmentSentientCodex.java
@@ -114,7 +114,10 @@ public class EnchantmentSentientCodex extends EnchantmentInsaneTweaksBase {
         int gid = Enchantment.getEnchantmentID(INSTANCE);
         for (int i = 0; i < list.tagCount(); i++) {
             NBTTagCompound en = list.getCompoundTagAt(i);
-            if (en.getShort("id") == gid) {
+            // getInteger on "id", not getShort: JustEnoughIDs widens enchantment ids past 32767 and
+            // rewrites vanilla's own reads, but not ours - and a truncated id here would silently
+            // report level 0. "lvl" stays a short; JEID only widens the id.
+            if (en.getInteger("id") == gid) {
                 return en.getShort("lvl");
             }
         }
